Added Standby/Charing State when plugged into USB
Processor will stay powered when plugged into USB (on next rev of PCB)
Added charging loop (in place of main loop) in this state
LCD is on to display battery percentage while charging.
A few bugs and fixes needed when transitioning between ON and standby states. (playing power on/off sounds appropriately, and ensuring 3.3V reg latching is handled properly)