[Feature] Add Web Debug
Added a web debug mode to run a WebSocket server on the Leaf, enabling a debug website to pull off stats and running the collection of measurements as quick as possible on the leaf.

This is useful for quick turnaround development of algorithms to give quick feedback to the Vario.