WiFi Over the air update (#21)
* Refactoring some of the Button and Page rendering to allow for a stack of modal page drawing

* Arduino support for etlcpp library

Not required for PlatformIO, but this adds support for Arduino IDE builds once the etlcpp library has been added.

Test Plan (using Arduino IDE):
```
Sketch uses 612976 bytes (46%) of program storage space. Maximum is 1310720 bytes.
Global variables use 91304 bytes (27%) of dynamic memory, leaving 236376 bytes for local variables. Maximum is 327680 bytes.
```

* Fixed build system failes.

Added a build flag to PlatformIO to know when we're building with that.

Added a new library to the Arduino github action so it can check out that library..

I still maintain that keeping this working in two build environments is making this project harder to maintain already :\

* - Added WiFi OTA and supporting tools

* Added QR Code library to Codebase.

QR Code library was not building with Arduino.  Followed the recommended approach to copy/paste it in and rename it.

It appears to be MIT license, so I think this is good.

* [Feature] WiFi Manager

* BugFix that may cause exception faults on two pages sharing no sub-menus
* Added WiFi Manager to be able to configure WiFi without the needed for an app under "Manual" mode.