[Refactor] KML Flight Log Refactoring. (#41)
* Drop Arduino Support

Arduino support means we can't re-factor code to be in sub-directories and is adding in too much friction.  As this project has outgrown it, let's drop it in favor of PlatformIO.

Will need issue to clean up documentation.

* [Refactor] KML Flight Log Refactoring.

Refactoring flight log recording to be in a class.  Moving KML logic out of top level files into their own class.  Work done in preperation for IGC files.

* Refactoring time to use local time

+  More efficient time formating