Changes "dev" build to allow Serial (#54)
* Changes "dev" build to allow Serial

- Dev mode disables USB Mass storage on boot
- Dev mode can "upload" without the need for the reset buttons now.
refactor: update USB mode and mass storage setup

Disable debug output for the SD card and adjust USB mode flags to 
ensure proper functionality during boot. Add a new endpoint to the 
web server for initiating mass storage mode. This improves the 
debugging experience and enhances the usability of the mass storage 
feature.

* [Work with Windows]

Work with windows when not behind a USB hub in release mode.